Add GitHub API v3 "List user repositories". fix #485 @1.0h
